API

HJB.CentralGradient

HJB.CentralWenoGradient

HJB.GradientMethod

HJB.Grid

HJB.Grid

HJB.LeftGradient

HJB.LeftWenoGradient

HJB.LocalLaxFriedrichsNHM

HJB.RightGradient

HJB.RightWenoGradient

HJB.SimpleNHM

HJB.StencilGradient

Base.step

HJB.DomainIndices

HJB.LeftBoundaryIndices

HJB.RightBoundaryIndices

HJB.allocate_grid

HJB.allocate_grid

HJB.getDomainAxes

HJB.gradient

HJB.hjb_plot

HJB.ind2state

HJB.state2ind

HJB.update_ghost_nodes!

HJB.weno

HJB.CentralGradient — Type

CentralGradient()

Returns second order central gradient,

\[\frac{u[i+1] - u[i-1]}{2 dx}\]

HJB.CentralWenoGradient — Type

CentralWenoGradient(ϵ=1e-6)

Returns the fifth order central WENO gradient, i.e., the average of the left and right Weno gradients.

HJB.GradientMethod — Type

GradientMethod

abstract type defining the method used to determine the spatial gradient of a grid. Given G <: GradientMethod, the general form is gradient(G, data, grid, ind, dim) to get the gradient of the array data at ind along dim dimension in a grid.

HJB.Grid — Type

Grid(x0, dx, Ns)

parameterize a grid, such that Array{T, D} can be mapped to a physical domain of D dimensions.

index (0, 0, ..., 0) will correspond to the physical point grid.x0

Note, this package makes heavy use of OffsetArrays to do its array indexing. Use ind2state and state2ind to correctly map between indices and physical points.

HJB.Grid — Method

Grid( (x1s, x2s, ...) )

a simpler constructor for a grid, where you pass in a tuple of xis <: AbstractRange.

HJB.LeftGradient — Type

LeftGradient()

Returns the first order left gradient,

\[\frac{u[i] - u[i-1]}{dx}\]

HJB.LeftWenoGradient — Type

LeftWenoGradient(ϵ=1e-6)

Returns the fifth order left WENO gradient. See Sec 3.4 of Osher, Fedkiw Level Set Methods and Dynamic Implicit Surfaces

HJB.LocalLaxFriedrichsNHM — Type

LocalLaxFriedrichsNHM(left_grad, right_grad, dissipation_func)

is a numerical hamiltonian method that implements

hamil(t, x, V, ∇V) = hamil(t, x, V, (gp + gm) / 2) - α' * (gp - gm) / 2

where gm, gp are the left and right gradients of V, and α = dissipation_func(t, x)

HJB.RightGradient — Type

RightGradient()

Returns first order right gradient

\[\frac{u[i+1] - u[i]}{dx} \]

HJB.RightWenoGradient — Type

RightWenoGradient(ϵ=1e-6)

Returns the fifth order right WENO gradient. See Sec 3.4 of Osher, Fedkiw Level Set Methods and Dynamic Implicit Surfaces

HJB.SimpleNHM — Type

SimpleNHM(left_grad, right_grad)

is a numerical hamiltonian method that implements

hamil(t, x, V, ∇V) = hamil(t, x, V, (gp + gm) / 2)

where gm, gp are the left and right gradients V.

HJB.StencilGradient — Type

StencilGradient(inds)

Returns a stencil based gradient operator, using Fornberg's method. Actually implemented using this trick from Steven Johnson. Given a list of relative inds, it can compute the weights w_i necessary such that the derivative is approximated by

\[\frac{1}{dx} \sum_{i \in inds} w[i] u[i]\]

Base.step — Method

step(ind, dim, n)

return the index of the cell n steps away from ind in the dim dimension.

HJB.DomainIndices — Method

DomainIndices(grid)

returns a CartesianIndices of a grid that correspond to the main computation domain, i.e., excluding any padding cells.

HJB.LeftBoundaryIndices — Method

LeftBoundaryIndices(grid, dim)

returns a CartesianIndices of a grid that correspond to the left boundary of the computation domain along the dim dimension

HJB.RightBoundaryIndices — Method

RightBoundaryIndices(grid, dim)

returns a CartesianIndices of a grid that correspond to the right boundary of the computation domain along the dim dimension

HJB.allocate_grid — Method

allocate_grid([T=Float64,] grid::Grid{D, F}, f, v=0) where {D, F}

returns an OffsetArray that will be used to store the value function. The OffsetArray will store values of type T, which is Float64 by default. The offset array is of D dimensions, and has the correct number of entries along each dimension. Each element of the array will be filled with f(x) where x is the physical location corresponding to the grid cell. All padding cells are filled with v.

HJB.allocate_grid — Method

allocate_grid([T=Float64,] grid::Grid{D, F}) where {D, F}

returns an OffsetArray that will be used to store the value function. The OffsetArray will store values of type T, which is Float64 by default. The offset array is of D dimensions, and has the correct number of entries along each dimension.

Elements are uninitialized.

HJB.getDomainAxes — Method

getDomainAxes(grid::Grid, dim)

returns a range that defines the axis of the grid along dim. Useful for plotting.

HJB.gradient — Method

gradient(data, grid, ind, dim)
gradient(method::GradientMethod, data, grid, ind, dim)

returns the gradient of data at ind along dimension dim using the grid and the method gradient method. Default method is CentralWenoGradient

HJB.hjb_plot — Function

hjb_plot(grid::G, data; idxs=(1,2), at=zeros(D), kwargs...) where {D, F, G<:Grid{D, F}}

A plotting utility to plot the solution of a hjb. Provide the grid, the data, and if the data is of dim > 2, also the dimensions you want to slice in, and the reference point at that you want to slice through.

Examples:

# create the grids
x1s = -1:0.025:1
x2s = -2:0.025:2
x3s = 0.0:0.5:2.0
grid1 = HJB.Grid((x1s,))
grid2 = HJB.Grid((x1s, x2s,))
grid3 = HJB.Grid((x1s, x2s, x3s))

# fill in some data
data1 = HJB.allocate_grid(grid1, x-> x[1]^2);
data2 = HJB.allocate_grid(grid2, x-> x[1]^2 + x[2]^2);
data3 = HJB.allocate_grid(grid3, x-> x[1]^2 + x[2]^2 - x[3]);

# plot 1D:
hjb_plot(grid1, data1)

# plot 2D:
hjb_plot(grid2, data2)

# plot 3D:
hjb_plot(grid3, data3, (1, 2), [0, 0, 1.0])

HJB.ind2state — Method

ind2state(grid::Grid, ind::CartesianIndex)

returns a SVector with the state at the ind in the grid

HJB.state2ind — Method

state2ind(grid::Grid, state)

returns a CartesianIndex corresponding to the state in the grid

HJB.update_ghost_nodes! — Method

update_ghost_nodes!(V, grid)

V is the array that contains the value function, specified using the grid.

updates (in-place) the values of the ghost nodes, using the boundary gradients and extrapolating. TODO(dev): support periodic boundaries

HJB.weno — Function

weno(v, ϵ=1e-6)

helper function to compute the WENO gradient given v=(v1,v2,v3,v4,v5) See Sec 3.5 of Osher, Fedkiw, Level Set Methods and Dynamic Implicit Surfaces.