Spatial Derivatives

Once we have a grid, we can take spatial derivatives using various methods.

The syntax is

  gradient(data, grid, ind, dim)
  gradient(method, data, grid, ind, dim)

which allows you determine the gradient of the data stored in data, based on the grid at a position ind along dimension dim. You can also specify the method to use. The mathematical expressions for each are listed in the docs below.

All of the gradient methods are sufficiently general to work in arbitrary dimensions, but assumes constant spacing of grid nodes.

Spatial Derivatives API

HJB.CentralGradient — Type

CentralGradient()

Returns second order central gradient,

\[\frac{u[i+1] - u[i-1]}{2 dx}\]

HJB.CentralWenoGradient — Type

CentralWenoGradient(ϵ=1e-6)

Returns the fifth order central WENO gradient, i.e., the average of the left and right Weno gradients.

HJB.GradientMethod — Type

GradientMethod

abstract type defining the method used to determine the spatial gradient of a grid. Given G <: GradientMethod, the general form is gradient(G, data, grid, ind, dim) to get the gradient of the array data at ind along dim dimension in a grid.

HJB.LeftGradient — Type

LeftGradient()

Returns the first order left gradient,

\[\frac{u[i] - u[i-1]}{dx}\]

HJB.LeftWenoGradient — Type

LeftWenoGradient(ϵ=1e-6)

Returns the fifth order left WENO gradient. See Sec 3.4 of Osher, Fedkiw Level Set Methods and Dynamic Implicit Surfaces

HJB.RightGradient — Type

RightGradient()

Returns first order right gradient

\[\frac{u[i+1] - u[i]}{dx} \]

HJB.RightWenoGradient — Type

RightWenoGradient(ϵ=1e-6)

Returns the fifth order right WENO gradient. See Sec 3.4 of Osher, Fedkiw Level Set Methods and Dynamic Implicit Surfaces

HJB.StencilGradient — Type

StencilGradient(inds)

Returns a stencil based gradient operator, using Fornberg's method. Actually implemented using this trick from Steven Johnson. Given a list of relative inds, it can compute the weights w_i necessary such that the derivative is approximated by

\[\frac{1}{dx} \sum_{i \in inds} w[i] u[i]\]

Base.step — Method

step(ind, dim, n)

return the index of the cell n steps away from ind in the dim dimension.

HJB.gradient — Method

gradient(data, grid, ind, dim)
gradient(method::GradientMethod, data, grid, ind, dim)

returns the gradient of data at ind along dimension dim using the grid and the method gradient method. Default method is CentralWenoGradient

HJB.weno — Function

weno(v, ϵ=1e-6)

helper function to compute the WENO gradient given v=(v1,v2,v3,v4,v5) See Sec 3.5 of Osher, Fedkiw, Level Set Methods and Dynamic Implicit Surfaces.